Refactoring testCameras
