+++ src/main/java/opencart/model/ProductPage.java
@@ -1,28 +1,22 @@
 package opencart.model;
 
+import io.qameta.allure.Step;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-public class ProductPage extends TopMenu {
+public class ProductPage extends BreadcrumbsMenu {
 
     @FindBy(css = "h1")
     private WebElement productName;
-
-    @FindBy(xpath = "//ul[@class='breadcrumb']")
-    private WebElement breadcrumbTop;
 
     protected ProductPage(WebDriver driver) {
         super(driver);
     }
 
+    @Step("Collect Actual Product Name Text.")
     public String getProductName() {
 
         return productName.getText();
-    }
-
-    public String getBreadcrumbTopText() {
-
-        return breadcrumbTop.getText();
     }
 }
