+++ src/main/java/opencart/model/SideMenu.java
@@ -2,7 +2,7 @@ package opencart.model;
 
 import org.openqa.selenium.WebDriver;
 
-public class SideMenu extends TopMenu {
+abstract class SideMenu extends TopMenu {
 
     protected SideMenu(WebDriver driver) {
         super(driver);
