+++ src/test/java/opencart/test/CamerasTest.java
